Implement fix for checkout process

+++ react/src/CheckoutComplete.jsx
@@ -1,41 +1,45 @@
 import React, { useEffect, useState } from "react";
+import { useParams } from "react-router-dom";
 import DisplayProduct from "./DisplayProduct";
 
 function Checkout() {
-  const [userId, setUserId] = useState("");
   const [products, setProducts] = useState([]);
   const [customHighTotal, setCustomHighTotal] = useState(0);
   const [enlarge, setEnlarge] = useState(false);
   const [enlargeProduct, setEnlargeProduct] = useState(false);
+  const { orderId } = useParams();
 
   useEffect(() => {
-    let oID = 0;
-    if (localStorage.getItem("oID")) {
-        oID = localStorage.getItem("oID");
-        localStorage.clear();
-    }
-    else if (userId) {
-        oID = 0;
-    }
-    fetch("/api/recentOrderDetails.php")
-      .then((response) => response.json())
-      .then((data) => {
-        setProducts(data);
-        for (let i = 0; i < data.length; ++i) {
-          if (data[i].product_id == 0) {
-            const temp = customHighTotal;
-            setCustomHighTotal(temp + (6 * data[i].product_quantity));
-          }
+    fetch("/api/checkout.php", {
+      method: "POST",
+      headers: { "Content-Type": "application/json" },
+      body: JSON.stringify({orderId}),
+    })
+    .then((response) => response.json())
+    .then((data) => {
+      if (data === 1) {
+        let oID = 0;
+        if (localStorage.getItem("oID")) {
+            oID = localStorage.getItem("oID");
+            localStorage.clear();
         }
-      });
-  }, []);
-
-  useEffect(() => {
-    fetch("/api/session.php")
-      .then((response) => response.json())
-      .then((data) => {
-        setUserId(data.userId);
-      });
+        fetch("/api/recentOrderDetails.php", {
+          method: "POST",
+          headers: { "Content-Type": "application/json" },
+          body: JSON.stringify({orderId}),
+        })
+        .then((response) => response.json())
+        .then((data) => {
+          setProducts(data);
+          for (let i = 0; i < data.length; ++i) {
+            if (data[i].product_id == 0) {
+              const temp = customHighTotal;
+              setCustomHighTotal(temp + (6 * data[i].product_quantity));
+            }
+          }
+        });
+      }
+    });
   }, []);
 
   const setPrice = (price, type, size) => {
@@ -76,15 +80,6 @@ function Checkout() {
       }
     }
     return price;
-  }
-
-  const determineDesign = (color) => {
-    if (color == 'Yellow' || color == 'Gray' || color == 'White') {
-      return ('customDesignBlack.png')
-    }
-    else {
-      return ('customDesign.png')
-    }
   }
 
   const handleOutsideClick = (event) => {
